broke log4j parsing out into separate module
